Added separate folder Auth for auth, SignIn, ForgotPassword, PrivatRoute
