+++ src/Components/Auth/ForgotPassword.js
